Merge develop into breaking.

+++ libsolutil/picosha2.h
@@ -95,10 +95,10 @@ void hash256_block(RaIter1 message_digest, RaIter2 first, RaIter2 last) {
     word_t w[64];
     std::fill(w, w + 64, 0);
     for (std::size_t i = 0; i < 16; ++i) {
-        w[i] = (static_cast<word_t>(mask_8bit(*(first + i * 4))) << 24) |
-               (static_cast<word_t>(mask_8bit(*(first + i * 4 + 1))) << 16) |
-               (static_cast<word_t>(mask_8bit(*(first + i * 4 + 2))) << 8) |
-               (static_cast<word_t>(mask_8bit(*(first + i * 4 + 3))));
+        w[i] = (static_cast<word_t>(mask_8bit(*(first + long(i) * 4))) << 24) |
+               (static_cast<word_t>(mask_8bit(*(first + long(i) * 4 + 1))) << 16) |
+               (static_cast<word_t>(mask_8bit(*(first + long(i) * 4 + 2))) << 8) |
+               (static_cast<word_t>(mask_8bit(*(first + long(i) * 4 + 3))));
     }
     for (std::size_t i = 16; i < 64; ++i) {
         w[i] = mask_32bit(ssig1(w[i - 2]) + w[i - 7] + ssig0(w[i - 15]) +
@@ -196,10 +196,10 @@ class hash256_one_by_one {
         std::copy(first, last, std::back_inserter(buffer_));
         std::size_t i = 0;
         for (; i + 64 <= buffer_.size(); i += 64) {
-            detail::hash256_block(h_, buffer_.begin() + i,
-                                  buffer_.begin() + i + 64);
+            detail::hash256_block(h_, buffer_.begin() + long(i),
+                                  buffer_.begin() + long(i) + 64);
         }
-        buffer_.erase(buffer_.begin(), buffer_.begin() + i);
+        buffer_.erase(buffer_.begin(), buffer_.begin() + long(i));
     }
 
     void finish() {
@@ -297,20 +297,20 @@ void hash256_impl(RaIter first, RaIter last, OutIter first2, OutIter last2, int,
 
 template <typename InputIter, typename OutIter>
 void hash256_impl(InputIter first, InputIter last, OutIter first2,
-                  OutIter last2, int buffer_size, std::input_iterator_tag) {
+                  OutIter last2, size_t buffer_size, std::input_iterator_tag) {
     std::vector<byte_t> buffer(buffer_size);
     hash256_one_by_one hasher;
     // hasher.init();
     while (first != last) {
-        int size = buffer_size;
-        for (int i = 0; i != buffer_size; ++i, ++first) {
+        size_t size = buffer_size;
+        for (size_t i = 0; i != buffer_size; ++i, ++first) {
             if (first == last) {
                 size = i;
                 break;
             }
             buffer[i] = *first;
         }
-        hasher.process(buffer.begin(), buffer.begin() + size);
+        hasher.process(buffer.begin(), buffer.begin() + ptrdiff_t(size));
     }
     hasher.finish();
     hasher.get_hash_bytes(first2, last2);
